requirements.txt
APScheduler==3.7.0
cassandra-driver==3.25.0
click==8.0.1
cmake==3.21.1.post1
colorama==0.4.4
colorhash==1.0.3
config==0.5.0.post0
configparser==5.0.2
cycler==0.10.0
Flask==1.1.4
Flask-Cors==3.0.10
Flask-MonitoringDashboard==3.1.0
geomet==0.2.1.post1
greenlet==1.1.1
imbalanced-learn==0.8.0
importlib-metadata==4.6.3
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==1.0.1
kiwisolver==1.3.1
kneed==0.7.0
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.19.5
pandas==1.1.5
Pillow==8.3.1
pip==21.2.4
psutil==5.8.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.1
setuptools==57.4.0
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.23
threadpoolctl==2.2.0
thrift==0.13.0
typing-extensions==3.10.0.0
tzlocal==2.1
Werkzeug==1.0.1
wheel==0.37.0
xgboost==1.4.2
zipp==3.5.0
intel-openmp==2021.3.0